schema: permissions BIGINT → INTEGER
We were already casting certain permissions to INTEGER in cogs/db, and
we don't really need 64 bits of permissions. 32 should be enough.

+++ schema.sql
@@ -24,14 +24,14 @@ CREATE TABLE IF NOT EXISTS revisions(
 
 CREATE TABLE IF NOT EXISTS role_permissions(
 	role BIGINT PRIMARY KEY,
-	permissions BIGINT NOT NULL);
+	permissions INTEGER NOT NULL);
 
 CREATE TABLE IF NOT EXISTS page_permissions(
 	page_id INTEGER NOT NULL REFERENCES pages ON DELETE CASCADE,
 	role BIGINT NOT NULL REFERENCES role_permissions ON DELETE CASCADE,
 	-- permissions to allow which overwrite role permissions
-	allow BIGINT NOT NULL DEFAULT 0,
+	allow INTEGER NOT NULL DEFAULT 0,
 	-- permissions to deny
-	deny BIGINT NOT NULL DEFAULT 0,
+	deny INTEGER NOT NULL DEFAULT 0,
 
 	PRIMARY KEY (page_id, role));
